migrate flask-jwt to flask-jwt-extended and integrate blocklist

+++ resources/item.py
@@ -1,5 +1,7 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required
+from flask_jwt_extended import get_jwt
+from flask_jwt_extended import get_jwt_identity
 from models.item import ItemModel
 
 
@@ -23,6 +25,7 @@ class Item(Resource):
             return item.json()
         return {"message": "item not found"}, 404
 
+    @jwt_required(fresh=True)
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {"message": "An item with name '{}' already exists.".format(name)}, 400
@@ -38,8 +41,12 @@ class Item(Resource):
 
         return item.json(), 201
 
-    @ jwt_required()
+    @jwt_required()
     def delete(self, name):
+        claims = get_jwt()
+        if not claims["is_admin"]:
+            return {"message": "Admin privilege required."}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -53,7 +60,7 @@ class Item(Resource):
         item = ItemModel.find_by_name(name)
 
         if item is None:
-            item - ItemModel(name, **data)
+            item = ItemModel(name, **data)
         else:
             item.price = data["price"]
 
@@ -63,5 +70,13 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_required(optional=True)
     def get(self):
-        return {"items": [item.json() for item in ItemModel.query.all()]}
+        user_id = get_jwt_identity()
+
+        items = [item.json() for item in ItemModel.find_all()]
+
+        if user_id:
+            return {"items": items}, 200
+        return {"items": [item["name"] for item in items],
+                "message": "More data available if you log in."}
